Change `AlgorithmLaunch` to `POST` on `Jobs`
