+++ main/schema/_generator.py
@@ -33,8 +33,6 @@ class CustomGenerator(SchemaGenerator):
                 'AffiliationUpdate': affiliation_update,
                 'Affiliation': affiliation,
                 'AlgorithmParameter': algorithm_parameter,
-                'AlgorithmLaunchSpec': algorithm_launch_spec,
-                'AlgorithmLaunch': algorithm_launch,
                 'Algorithm': algorithm,
                 'AlgorithmSpec': algorithm_spec,
                 'AlgorithmManifest': algorithm_manifest,
@@ -92,6 +90,7 @@ class CustomGenerator(SchemaGenerator):
                 'Invitation': invitation,
                 'JobNode': job_node,
                 'Job': job,
+                'JobSpec': job_spec,
                 'JobCluster': job_cluster,
                 'JobClusterSpec': job_cluster_spec,
                 'LeafTypeSpec': leaf_type_spec,
